Fix waitTime overflow
Since micros() overflows approximately every 70 minutes, and all waits are
with respect to bitStartTime anyway, only store the wait duration and
not the wake-up time. This solves all overflow issues.

+++ CECWire.h
@@ -71,7 +71,7 @@ private:
 
 	bool _lastLineState;
 	unsigned long _bitStartTime;
-	unsigned long _waitTime;
+	unsigned int _waitTime;
 
 	int _xmitretry;
 	enum {
